fix: feature-detection for chrome.sockets
chrome-net patch:
https://github.com/lidel/chrome-net/commit/838ffde4a33721888f74783821e0486dfcc88797

  chrome-net did not check if chrome.sockets.* actually exist.
  This caused problems with extensions built for Chromium-based browsers
  that tried to maintain the single codebase and do feature-detection
  at runtime.

feature detection for embedded js-ipfs:

  If chrome.sockets is available, embedded-brave.js will be used
  instead of regular embedded.js

notify user about socket errors:

  If port is already taken, a notification is displayed.
  We will most likely improve this soon (eg. by finding the next free port)
  but is good enough for now.
