+++ package.json
@@ -107,10 +107,11 @@
   },
   "dependencies": {
     "@material/switch": "1.0.1",
+    "@nodeutils/defaults-deep": "1.1.0",
     "browser-process-hrtime": "1.0.0",
     "choo": "6.13.1",
     "chrome-dgram": "3.0.1",
-    "chrome-net": "3.3.1",
+    "chrome-net": "https://github.com/lidel/chrome-net/tarball/838ffde4a33721888f74783821e0486dfcc88797/chrome-net.tar.gz",
     "doc-sniff": "1.0.1",
     "drag-and-drop-files": "0.0.1",
     "file-type": "10.9.0",
